feat: restructure codebase into ISO 13374 sub-packages + architecture docs
Reorganize Phase 1 modules into ISO 13374-aligned sub-packages:
- signal_acquisition/ (Block 1): signal loading + LRU repository
- signal_processing/ (Block 2): PSD, STFT, envelope spectrum
- diagnostics/ (Blocks 3-4): bearing faults, catalog, ISO 10816
- decision_support/ (Block 6): integrated diagnosis pipeline
- prognostics/ (Block 5): placeholder for Phase 2

Backward compatibility maintained via shim modules in src/ root.
Both old and new import paths work:
  from predictive_maintenance_mcp.spectral import compute_psd  # old
  from predictive_maintenance_mcp.signal_processing import compute_psd  # new

Add architecture guide (docs/architecture.md) and Sphinx API docs setup.
All 225 tests pass with zero regressions.
