+++ docs/api/decision_support.rst
@@ -1,0 +1,8 @@
+Decision Support (ISO 13374 Block 6)
+=====================================
+
+Integrated diagnosis pipeline.
+
+.. automodule:: predictive_maintenance_mcp.decision_support.diagnosis_pipeline
+   :members:
+   :undoc-members:
