+++ docs/api/diagnostics.rst
@@ -1,0 +1,13 @@
+Diagnostics (ISO 13374 Blocks 3-4)
+====================================
+
+Bearing fault detection and severity assessment.
+
+.. automodule:: predictive_maintenance_mcp.diagnostics.bearing_analyzer
+   :members:
+
+.. automodule:: predictive_maintenance_mcp.diagnostics.bearing_catalog
+   :members:
+
+.. automodule:: predictive_maintenance_mcp.diagnostics.iso10816
+   :members:
