+++ docs/api/signal_acquisition.rst
@@ -1,0 +1,12 @@
+Signal Acquisition (ISO 13374 Block 1)
+=======================================
+
+Signal loading and in-memory caching.
+
+.. automodule:: predictive_maintenance_mcp.signal_loader
+   :members:
+   :undoc-members:
+
+.. automodule:: predictive_maintenance_mcp.signal_repository
+   :members:
+   :undoc-members:
