+++ docs/api/signal_processing.rst
@@ -1,0 +1,7 @@
+Signal Processing (ISO 13374 Block 2)
+======================================
+
+Spectral analysis: PSD, STFT, envelope spectrum.
+
+.. automodule:: predictive_maintenance_mcp.signal_processing.spectral
+   :members:
