+++ docs/index.rst
@@ -1,0 +1,20 @@
+predictive-maintenance-mcp
+==========================
+
+Industrial-grade predictive maintenance and fault diagnosis via Model Context Protocol.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents
+
+   architecture
+   api/signal_acquisition
+   api/signal_processing
+   api/diagnostics
+   api/decision_support
+
+Indices
+-------
+
+* :ref:`genindex`
+* :ref:`modindex`
