+++ src/decision_support/__init__.py
@@ -1,0 +1,9 @@
+"""
+ISO 13374 Block 6 — Decision Support (Advisory Generation).
+
+Integrated diagnosis pipeline combining spectral, bearing, and severity analyses.
+"""
+
+from .diagnosis_pipeline import (  # noqa: F401
+    diagnose_vibration,
+)
