+++ src/diagnostics/__init__.py
@@ -1,0 +1,19 @@
+"""
+ISO 13374 Blocks 3-4 — State Detection & Health Assessment.
+
+Bearing fault detection, catalog lookup, ISO 10816/20816 severity.
+"""
+
+from .bearing_analyzer import (  # noqa: F401
+    check_bearing_fault_peak,
+    check_all_bearing_faults,
+    lookup_bearing_and_compute,
+)
+from .bearing_catalog import (  # noqa: F401
+    lookup_bearing,
+    compute_fault_frequencies,
+    list_catalog_bearings,
+)
+from .iso10816 import (  # noqa: F401
+    assess_vibration_severity,
+)
