+++ src/prognostics/__init__.py
@@ -1,0 +1,5 @@
+"""
+ISO 13374 Block 5 — Prognostic Assessment.
+
+Placeholder for Phase 2: RUL estimation, trend analysis.
+"""
