+++ src/signal_acquisition/loaders.py
@@ -1,0 +1,3 @@
+"""ISO 13374 Block 1 — Signal loaders (delegates to signal_loader module)."""
+from ..signal_loader import *  # noqa: F401,F403
+from ..signal_loader import SUPPORTED_EXTENSIONS  # noqa: F401
