+++ src/signal_acquisition/repository.py
@@ -1,0 +1,3 @@
+"""ISO 13374 Block 1 — Signal repository (delegates to signal_repository module)."""
+from ..signal_repository import *  # noqa: F401,F403
+from ..signal_repository import SignalRepository, get_repository  # noqa: F401
