+++ src/signal_processing/__init__.py
@@ -1,0 +1,11 @@
+"""
+ISO 13374 Block 2 — Signal Processing.
+
+Spectral analysis: PSD (Welch), STFT spectrogram, envelope spectrum.
+"""
+
+from .spectral import (  # noqa: F401
+    compute_psd,
+    compute_stft_spectrogram,
+    compute_envelope_spectrum,
+)
